Adding pristine defaults.
