+++ keyboards/gggw/crosses/3x5/keymaps/stock/rules.mk
