+++ keyboards/gggw/crosses/3x6/keymaps/stock/rules.mk
